requirements.txt
aiogram==3.4.1
mega.py==1.0.8
